Move Card class to its own file
