+++ Assets/Scripts/Core/Card.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 25434a7b582c7f2419e9849c12898f7d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
